requirements.txt
requests>=2.28.0
google-generativeai>=0.5.0
beautifulsoup4>=4.12.0
lxml>=5.1.0